% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovakun_internal.R
\name{anova.modeler}
\alias{anova.modeler}
\title{Default method for anova.modeler}
\usage{
\method{anova}{modeler}(
  dat,
  design,
  factnames = NA,
  type2 = FALSE,
  dmat = NULL,
  flev = NULL,
  cellN = NULL,
  full.elem = NA,
  epsi.effect = NA,
  lb = FALSE,
  gg = FALSE,
  hf = FALSE,
  cm = FALSE,
  auto = FALSE,
  autov = NULL,
  mau = FALSE,
  har = FALSE,
  iga = FALSE,
  ciga = FALSE,
  eta = FALSE,
  peta = FALSE,
  geta = NA,
  eps = FALSE,
  peps = FALSE,
  geps = NA,
  omega = FALSE,
  omegana = FALSE,
  pomega = FALSE,
  gomega = NA,
  gomegana = NA,
  prep = FALSE,
  nesci = FALSE,
  inter = NA,
  bet.mse = NULL,
  gss.qT = NA,
  post.esdenomis = NA,
  post.df.adj = NA,
  post.mse.adj = NA,
  esboot = FALSE,
  es.conf.level = 0.95
)
}
\description{
Default method for anova.modeler
}
\keyword{internal}
