% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovakun_internal.R
\name{anova.output}
\alias{anova.output}
\title{S3 generic for anova.output}
\usage{
\method{anova}{output}(maxfact, exe.info, baseresults, mainresults, postresults = NA, sep = " ")
}
\description{
S3 generic for anova.output
}
