% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.r
\docType{data}
\name{data_fict_between}
\alias{data_fict_between}
\title{Example dataset: data_fict_between}
\format{
A data frame with 16 rows and 4 variables:
\describe{
\item{participant}{Participant ID}
\item{group}{Group (\code{"control"} or \code{"intervention"})}
\item{gender}{Gender (\code{"woman"} or \code{"man"})}
\item{n_correct}{Number of correct responses}
}
}
\usage{
data(data_fict_between)
}
\description{
This is a fictitious dataset for 2*2 between-participants ANOVA.
}
\keyword{datasets}
