% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.r
\docType{data}
\name{data_fict_mixed}
\alias{data_fict_mixed}
\title{Example dataset: data_fict_mixed}
\format{
A data frame with 16 rows and 4 variables:
\describe{
\item{participant}{Participant ID}
\item{gender}{Gender (\code{"woman"} or \code{"man"})}
\item{condition}{Condition (\code{"control"} or \code{"experimental"})}
\item{rt}{Mean response time (in milliseconds)}
}
}
\usage{
data(data_fict_mixed)
}
\description{
This is a fictitious dataset for 2*2 mixed-design ANOVA.
}
\keyword{datasets}
