% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.r
\docType{data}
\name{data_fict_within}
\alias{data_fict_within}
\title{Example dataset: data_fict_within}
\format{
A data frame with 64 rows and 4 variables:
\describe{
\item{participant}{Participant ID}
\item{condition}{Condition (\code{"control"} or \code{"experimental"})}
\item{time}{Time (\code{"pre"} or \code{"post"})}
\item{score}{Test score}
}
}
\usage{
data(data_fict_within)
}
\description{
This is a fictitious dataset for 2*2 within-participants ANOVA.
}
\keyword{datasets}
