% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.r
\docType{data}
\name{data_snakemr}
\alias{data_snakemr}
\title{Example dataset: data_snakemr}
\format{
A data frame with 480 rows and 5 variables:
\describe{
\item{participant}{Participant ID}
\item{angle}{Angular disparity between two Stimuli (in degrees)}
\item{shape}{Stimulus shape (\code{"human"} or \code{"snake"})}
\item{face}{Presence of face (\code{"absent"} or \code{"present"})}
\item{rt}{Mean response time (in seconds)}
}
}
\source{
Muto, H., & Nagai, M. (2020). Mental rotation of cubes with a snake face: The role of the human-body analogy revisited. \emph{Visual Cognition, 28}(2), 106–111. \url{https://doi.org/10.1080/13506285.2020.1727598}
}
\usage{
data(data_snakemr)
}
\description{
This dataset was used in a study on mental rotation involving cube stimuli with snake-like faces.
}
\keyword{datasets}
