% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree_to_radian.r
\name{degree_to_radian}
\alias{degree_to_radian}
\title{Convert degrees to radians}
\usage{
degree_to_radian(x)
}
\arguments{
\item{x}{A numeric vector representing an angle in degrees.}
}
\value{
A numeric vector giving the angle in radians.
}
\description{
This function converts an angle measured in degrees to radians.
}
\examples{
degree_to_radian(180) # pi
degree_to_radian(90)  # pi/2

}
