% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdi.r
\name{hdi}
\alias{hdi}
\title{Compute the Highest Density Interval (HDI)}
\usage{
hdi(x, prob = 0.95)
}
\arguments{
\item{x}{A numeric vector of data values for which the HDI will be calculated.}

\item{prob}{The probability mass of the HDI to calculate. Default is 0.95.}
}
\value{
A numeric vector of length 2 representing the lower and upper bounds of the HDI.
}
\description{
This function calculates the highest density interval (HDI) for a given
set of samples. The HDI is the interval that contains a specified
proportion (default 95\%) of the highest density values from the distribution.
}
\details{
The function first sorts the input data and computes the necessary quantiles using inear interpolation.
It then identifies the interval that contains the highest density, ensuring that the HDI encompasses the specified probability mass.
Note that this function does not support the case where two or more separate
HDIs exist.
}
\examples{
set.seed(123)
data <- rnorm(1000)  # Generate random normal data
hdi(data, prob = 0.95)

}
