% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaded_packages_version.r
\name{loaded_packages_version}
\alias{loaded_packages_version}
\title{Get a list of currently loaded packages and their versions}
\usage{
loaded_packages_version()
}
\value{
A data frame with two columns:
\describe{
\item{package}{The name of the loaded package.}
\item{version}{The version of the loaded package.}
}
}
\description{
This function retrieves the names and versions of all currently loaded R packages.
It returns a data frame with the package names in the first column and their respective
versions in the second column.
}
\examples{
# Get the list of loaded packages with versions
loaded_packages_version()

}
