% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_stat.r
\name{mode_stat}
\alias{mode_stat}
\title{Compute the Mode of a Distribution}
\usage{
mode_stat(z, n = 512)
}
\arguments{
\item{z}{A numeric vector containing the values from which the mode is to be estimated.}

\item{n}{An integer specifying the number of equally spaced grid points for kernel density estimation (default is 512).}
}
\value{
A numeric value representing the estimated mode of the distribution.
}
\description{
This function estimates the mode (the most frequent value) of a given
numeric vector using kernel density estimation. The mode is identified
as the value where the kernel density is maximized.
}
\examples{
# Generate random data from a normal distribution
set.seed(123)
data <- rnorm(1000)

# Compute the mode of the distribution
mode_stat(data)

}
