% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovakun_internal.R
\docType{package}
\name{mutolabr-package}
\alias{mutolabr}
\alias{mutolabr-package}
\title{Internal Helper Functions for anovakun}
\description{
This file contains internal functions used within the anovakun package.
These functions are not exported and should not be used directly by end users.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mutopsy/mutolabr}
}

}
\author{
\strong{Maintainer}: Hiroyuki Muto \email{mutopsy@omu.ac.jp}

}
\keyword{internal}
