% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdir.r
\name{pdir}
\alias{pdir}
\title{Probability of Direction (pd) Calculation}
\usage{
pdir(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be removed before calculation.
Default is \code{FALSE}.}
}
\value{
A numeric value between 0 and 1, representing the Probability of Direction (Pd).
The value is calculated as the proportion of positive values in \code{x}, or 1 minus
the proportion if it is less than 0.5.
}
\description{
This function calculates the probability of direction (pd), which is the proportion of
the distribution of the same sign as its median’s and varies from 0.50 to 1.00.
This is an index of effect existence, proposed by Makowski et al. (2019).
}
\examples{
set.seed(123)
x <- rnorm(1000)
pdir(x)  # Returns the Probability of Direction for x
pdir(x, na.rm = TRUE)  # Removes NA values before calculation

}
\references{
Makowski, D., Ben-Shachar, M. S., Chen, S. H. A., & Lüdecke, D. (2019). Indices of effect
existence and significance in the Bayesian framework. Frontiers in Psychology, 10, Article
2767. https://doi.org/10.3389/fpsyg.2019.02767
}
