% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radian_to_degree.r
\name{radian_to_degree}
\alias{radian_to_degree}
\title{Convert radians to degrees}
\usage{
radian_to_degree(x)
}
\arguments{
\item{x}{A numeric vector representing an angle in radians.}
}
\value{
A numeric vector giving the angle in degrees.
}
\description{
This function converts an angle measured in radians to degrees.
}
\examples{
radian_to_degree(pi)     # 180
radian_to_degree(pi / 2) # 90

}
