% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gg0.r
\name{theme_gg0}
\alias{theme_gg0}
\title{A Default ggplot2 Theme}
\usage{
theme_gg0()
}
\value{
A \code{ggplot2} theme object.
}
\description{
Returns a simplified ggplot2 theme based on \code{theme_bw()}, with grid lines removed,
black axis text, and no legend title. This theme is useful for producing cleaner, publication-ready plots.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  theme_gg0()
}
