% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_desc.r
\name{var_desc}
\alias{var_desc}
\title{Compute Population Variance}
\usage{
var_desc(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector containing the data.}

\item{na.rm}{A logical value indicating whether to remove missing values (\code{NA}) before computation. Default is \code{FALSE}.}
}
\value{
A numeric value representing the population variance of \code{x}.
}
\description{
This function calculates the population variance of a numeric vector.
Unlike the default \code{var()} function in R, which computes the sample variance,
this function adjusts the calculation to use $$n$$ instead of $$n-1$$ in the denominator.
}
\examples{
data <- c(10, 20, 30, 40, 50, NA)
var_desc(data)          # Compute variance including NA (returns NA)
var_desc(data, na.rm = TRUE)  # Compute variance ignoring NA

}
\seealso{
\code{\link{sd_desc}} for computing the population standard deviation.
}
