% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdir.r
\name{pdir}
\alias{pdir}
\title{Compute the Highest Density Interval (HDI)}
\usage{
pdir(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector of samples from the distribution.}

\item{prob}{A numeric value between 0 and 1 specifying the probability mass
for the HDI (default is 0.95).}
}
\value{
A numeric vector containing the values within the HDI.
}
\description{
This function calculates the highest density interval (HDI) for a given
set of samples. The HDI is the interval that contains a specified
proportion (default 95\%) of the highest density values from the distribution.
Note that this function does not support the case where two or more separate
HDIs exist.
}
\examples{
# Generate some random posterior samples
set.seed(123)
samples <- rnorm(1000, mean = 0, sd = 1)

# Compute the 95\% HDI
hdi(samples, prob = 0.95)

}
