% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualangle.r
\name{visualangle}
\alias{visualangle}
\title{Compute visual angle or stimulus size in pixels}
\usage{
visualangle(
  distance_cm,
  monitor_cm,
  monitor_px,
  stim_px = NULL,
  stim_angle_deg = NULL
)
}
\arguments{
\item{distance_cm}{Viewing distance (cm).}

\item{monitor_cm}{Physical size of the monitor (cm) along the relevant
dimension (e.g., width).}

\item{monitor_px}{Number of pixels along the same dimension as
\code{monitor_cm}.}

\item{stim_px}{Stimulus size in pixels. Set to \code{NULL} if you want to
compute \code{stim_px} from \code{stim_angle_deg}.}

\item{stim_angle_deg}{Stimulus size in degrees. Set to \code{NULL} if you
want to compute \code{stim_angle_deg} from \code{stim_px}.}
}
\value{
If \code{stim_px} is provided: a numeric vector of visual angles (degrees). \cr
If \code{stim_angle_deg} is provided: a numeric vector of stimulus sizes (pixels).
}
\description{
If \code{stim_px} is provided and \code{stim_angle_deg} is \code{NULL},
the function returns the visual angle of the stimulus in degrees.
If \code{stim_angle_deg} is provided and \code{stim_px} is \code{NULL},
the function returns the stimulus size in pixels that subtends the
specified visual angle.
}
\details{
Exactly one of \code{stim_px} and \code{stim_angle_deg} must be non-NULL.
}
\examples{
# 1) px -> degree
visualangle(
  distance_cm = 60,
  monitor_cm = 53.1,
  monitor_px = 1920,
  stim_px = 200
)

# 2) degree -> px
visualangle(
  distance_cm = 60,
  monitor_cm = 53.1,
  monitor_px = 1920,
  stim_angle_deg = 5
)

}
